# cipher enum bug fixes and improvements