--- a/pkg/tlsx/auto/util.go
+++ b/pkg/tlsx/auto/util.go
@@ -1,6 +1,7 @@
 package auto
 
 import (
+	"github.com/projectdiscovery/tlsx/pkg/tlsx/openssl"
 	"github.com/projectdiscovery/tlsx/pkg/tlsx/tls"
 	"github.com/projectdiscovery/tlsx/pkg/tlsx/ztls"
 	sliceutil "github.com/projectdiscovery/utils/slice"
@@ -13,7 +14,9 @@ var (
 
 func init() {
 	allCiphersNames = append(tls.AllCiphersNames, ztls.AllCiphersNames...)
+	allCiphersNames = append(allCiphersNames, openssl.AllCiphersNames...)
 	supportedTlsVersions = append(tls.SupportedTlsVersions, ztls.SupportedTlsVersions...)
+	supportedTlsVersions = append(supportedTlsVersions, openssl.SupportedTLSVersions...)
 	allCiphersNames = sliceutil.Dedupe(allCiphersNames)
 	supportedTlsVersions = sliceutil.Dedupe(supportedTlsVersions)
 }