# Proposed Changes

- TLS Version enumeration is enabled internally when using `-cipher-enum` flag
- Print Cipher Enum results in standard CLI mode (Only shown in JSON output now)
- Bug fix Missing OpenSSL ciphers in auto mode
-  Cipher Enum Optimization
   - Skip cipher enumeration in ctls mode when version is TLS 1.3 which caused false positives
   - Adds EnumMode and skip unsupported ciphers by client in Enum Mode

closes #167  #164